renamed init file to blue_init
