+++ wsgi.py
@@ -1,4 +1,4 @@
-from . import app
+from blue_init import app
 
 if __name__ == "__main__":
     app.run()
